TimelineZoom is a float

+++ src/Game/Osu/OszLoader/OsuParser/Editor.hs
@@ -33,5 +33,5 @@ beatDivisorP = "BeatDivisor: " *> decimal
 gridSizeP ∷ Parser Int
 gridSizeP = "GridSize: " *> decimal
 
-timelineZoomP ∷ Parser Int
-timelineZoomP = "TimelineZoom: " *> decimal
+timelineZoomP ∷ Parser Double
+timelineZoomP = "TimelineZoom: " *> double
